corrections

+++ libft/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: denizozd <denizozd@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/10 14:06:11 by denizozd          #+#    #+#             */
-/*   Updated: 2023/11/10 14:28:46 by denizozd         ###   ########.fr       */
+/*   Updated: 2023/11/16 16:53:42 by denizozd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlcpy(char *dst, const char *src, size_t size)
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	size_t	i;
